Seperate Session Workflow from being inside InformationView
