--- Misa.Ui.Avalonia/Features/Details/Information/Extensions/DescriptionView.axaml.cs
@@ -1,13 +1,0 @@
-﻿using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace Misa.Ui.Avalonia.Features.Details.Information.Extensions;
-
-public partial class DescriptionView : UserControl
-{
-    public DescriptionView()
-    {
-        InitializeComponent();
-    }
-}
