+++ Misa.Ui.Avalonia/Features/Details/Information/Extensions/Descriptions/DescriptionView.axaml
@@ -3,9 +3,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="Misa.Ui.Avalonia.Features.Details.Information.Extensions.DescriptionView"
-             xmlns:parent="clr-namespace:Misa.Ui.Avalonia.Features.Details.Information.Extensions"
-             x:DataType="parent:DescriptionViewModel">
+             x:Class="Misa.Ui.Avalonia.Features.Details.Information.Extensions.Descriptions.DescriptionView"
+             xmlns:descriptions="clr-namespace:Misa.Ui.Avalonia.Features.Details.Information.Extensions.Descriptions"
+             x:DataType="descriptions:DescriptionViewModel">
     
         <Border Padding="12"
             Margin="10"
