+++ Misa.Ui.Avalonia/Features/Details/Information/Extensions/Descriptions/DescriptionView.axaml.cs
@@ -1,0 +1,11 @@
+﻿using Avalonia.Controls;
+
+namespace Misa.Ui.Avalonia.Features.Details.Information.Extensions.Descriptions;
+
+public partial class DescriptionView : UserControl
+{
+    public DescriptionView()
+    {
+        InitializeComponent();
+    }
+}
