+++ Misa.Ui.Avalonia/Features/Details/Information/Extensions/Sessions/PauseSessionView.axaml
@@ -1,0 +1,50 @@
+﻿<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="Misa.Ui.Avalonia.Features.Details.Information.Extensions.Sessions.PauseSessionView"
+             xmlns:information="clr-namespace:Misa.Ui.Avalonia.Features.Details.Information.Extensions.Sessions"
+             x:DataType="information:PauseSessionViewModel">
+    
+    
+    <Border Padding="10"
+                        CornerRadius="10"
+                        BorderThickness="1"
+                        BorderBrush="{StaticResource BrushBorder}"
+                        Background="{StaticResource BrushPrimaryBackground}">
+                    <StackPanel Spacing="10">
+        
+                        <TextBlock Text="Pause session"
+                                   FontWeight="SemiBold"/>
+        
+                        <Grid ColumnDefinitions="*"
+                              RowDefinitions="Auto,Auto"
+                              RowSpacing="8"
+                              ColumnSpacing="10">
+        
+                            <TextBlock Grid.Row="0" VerticalAlignment="Top" Text="What is the reason to pause your session?" />
+                            <StackPanel Grid.Row="1" Spacing="8">
+                                <TextBox Classes="misa-input"
+                                         Text="{Binding PauseReason, Mode=TwoWay}"
+                                         AcceptsReturn="True"
+                                         TextWrapping="Wrap"
+                                         MinHeight="80"/>
+        
+                                <StackPanel Orientation="Horizontal"
+                                            Spacing="8"
+                                            HorizontalAlignment="Right">
+                                    <Button Classes="misa-action"
+                                            Content="Cancel"/>
+                                    <Button Classes="misa-action"
+                                            Content="Save"
+                                            Command="{Binding PauseSessionCommand}" />
+                                </StackPanel>
+                            </StackPanel>
+        
+                        </Grid>
+        
+                    </StackPanel>
+                </Border>
+    
+</UserControl>
