+++ Misa.Ui.Avalonia/Features/Details/Information/Extensions/Sessions/PauseSessionView.axaml.cs
@@ -1,0 +1,13 @@
+﻿using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace Misa.Ui.Avalonia.Features.Details.Information.Extensions.Sessions;
+
+public partial class PauseSessionView : UserControl
+{
+    public PauseSessionView()
+    {
+        InitializeComponent();
+    }
+}
