+++ Misa.Ui.Avalonia/Features/Details/Information/Extensions/Sessions/SessionContinueView.axaml
@@ -1,0 +1,10 @@
+﻿<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="Misa.Ui.Avalonia.Features.Details.Information.Extensions.Sessions.SessionContinueView"
+             xmlns:information="clr-namespace:Misa.Ui.Avalonia.Features.Details.Information.Extensions.Sessions"
+             x:DataType="information:SessionContinueViewModel">
+    Welcome to Avalonia!
+</UserControl>
