+++ Misa.Ui.Avalonia/Features/Details/Information/Extensions/Sessions/SessionContinueViewModel.cs
@@ -1,0 +1,6 @@
+﻿namespace Misa.Ui.Avalonia.Features.Details.Information.Extensions.Sessions;
+
+public class SessionContinueViewModel
+{
+    
+}
