+++ Misa.Ui.Avalonia/Features/Details/Information/Extensions/Sessions/StopSessionView.axaml
@@ -1,0 +1,66 @@
+﻿<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="Misa.Ui.Avalonia.Features.Details.Information.Extensions.Sessions.StopSessionView"
+             xmlns:information="clr-namespace:Misa.Ui.Avalonia.Features.Details.Information.Extensions.Sessions"
+             x:DataType="information:StopSessionViewModel">
+    
+    <Border Padding="10"
+            CornerRadius="10"
+            BorderThickness="1"
+            BorderBrush="{StaticResource BrushBorder}"
+            Background="{StaticResource BrushPrimaryBackground}">
+        <StackPanel Spacing="10">
+
+            <TextBlock Text="Stop session"
+                       FontWeight="SemiBold"/>
+
+            <Grid ColumnDefinitions="140,*"
+                  RowDefinitions="Auto,Auto,Auto"
+                  RowSpacing="8"
+                  ColumnSpacing="10">
+
+                <TextBlock Grid.Row="0" Grid.Column="0" VerticalAlignment="Center" Text="Efficiency" />
+                <ComboBox Grid.Row="0" Grid.Column="1"
+                          Classes="misa-input"
+                          ItemsSource="{Binding EfficiencyTypes}"
+                          SelectedValue="{Binding EfficiencyId}"
+                          SelectedValueBinding="{Binding Id}"
+                          DisplayMemberBinding="{Binding Name}"/>
+
+                <TextBlock Grid.Row="1" Grid.Column="0" VerticalAlignment="Center" Text="Concentration" />
+                <ComboBox Grid.Row="1" Grid.Column="1"
+                          Classes="misa-input"
+                          ItemsSource="{Binding ConcentrationTypes}"
+                          SelectedValue="{Binding ConcentrationId}"
+                          SelectedValueBinding="{Binding Id}"
+                          DisplayMemberBinding="{Binding Name}" />
+
+                <TextBlock Grid.Row="2" Grid.Column="0" VerticalAlignment="Top" Text="Summary" />
+                <StackPanel Grid.Row="2" Grid.Column="1" Spacing="8">
+                    <TextBox Classes="misa-input"
+                             Text="{Binding Summary, Mode=TwoWay}"
+                             AcceptsReturn="True"
+                             TextWrapping="Wrap"
+                             MinHeight="80"
+                             Watermark="What happened? Short summary." />
+
+                    <StackPanel Orientation="Horizontal"
+                                Spacing="8"
+                                HorizontalAlignment="Right">
+                        <Button Classes="misa-action"
+                                Content="Cancel"/>
+                        <Button Classes="misa-action"
+                                Content="Save"
+                                Command="{Binding StopSessionCommand}" />
+                    </StackPanel>
+                </StackPanel>
+
+            </Grid>
+
+        </StackPanel>
+    </Border>
+
+</UserControl>
